+++ main/include/audio_process.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void audio_process_test();